[STINS-038] implemented-results-page
Issue: We did not have a result page before so we were just replacing the place holder with the content that needed to be on the result page.

Solution: Uploaded the content that was needed on the result page
Risk: yes, it will help the people implementing the email input section on that page.

Reviewed by: Jason Ko, Jason Zhao, Aidan Dayvyd